fix: use chakra tab
